+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added `--json` flag for `sparv modules` command.
 - The command `sparv modules` now also takes annotators in addition to modules as arguments to filter the output.
 - The detailed `sparv modules` listings now include info about accepted values for config variables.
+- Sparv now respects the `JAVA_HOME` environment variable.
 
 ### Changed